+++ main.go
@@ -57,7 +57,7 @@ func main() {
 	}
 	app := cli.NewApp()
 	app.Name = "ezb_srv"
-	app.Version = "0.1.2"
+	app.Version = "0.1.3"
 	app.Usage = "ezBastion frontend server."
 
 	app.Commands = []cli.Command{