+++ middleware/routeParser.go
@@ -36,6 +36,8 @@ func RouteParser(c *gin.Context) {
 		"middleware": "RouteParser",
 		"xtrack":     trace.Xtrack,
 	})
+	logg.Debug("start")
+
 	escapedPath := c.Request.URL.EscapedPath()
 	path := s.Split(escapedPath, "/")
 	// fmt.Printf("escapedPath: '%d'\n", len(path))
@@ -85,13 +87,13 @@ func RouteParser(c *gin.Context) {
 		var CurrentAction []models.ViewApi
 		for i := range ViewApis {
 			if ViewApis[i].Ctrlver == version {
-				// fmt.Println(version)
+				logg.Debug("ViewApis version match:", version)
 				if ViewApis[i].Ctrl == ctrl {
-					// fmt.Println(ctrl)
+					logg.Debug("ViewApis ctrl match:", ctrl)
 					if ViewApis[i].Action == action {
-						// fmt.Println(action)
+						logg.Debug("ViewApis action match:", action)
 						if s.ToUpper(ViewApis[i].Access) == access {
-							// fmt.Println(c.Request.Method)
+							logg.Debug("ViewApis access match:", access)
 							apiFound = true
 							// CurrentAction = ViewApis[i]
 							CurrentAction = append(CurrentAction, ViewApis[i])