+++ middleware/store.go
@@ -36,6 +36,8 @@ func Store(storage cache.Storage, conf *models.Configuration) gin.HandlerFunc {
 			"middleware": "store",
 			"xtrack":     trace.Xtrack,
 		})
+		logg.Debug("start")
+
 		routeType, _ := c.MustGet("routeType").(string)
 		if routeType == "worker" {
 			usr, _ := c.Get("account")
@@ -46,6 +48,7 @@ func Store(storage cache.Storage, conf *models.Configuration) gin.HandlerFunc {
 				c.AbortWithError(http.StatusInternalServerError, errors.New("#S0001"))
 				return
 			}
+			logg.Debug("found ", len(viewApi), " viewApi for user ", account.Name)
 			if len(viewApi) == 0 {
 				logg.Error("NO API FOUND FOR ", account.Name)
 				c.AbortWithError(http.StatusForbidden, errors.New("#S0002"))
@@ -59,6 +62,7 @@ func Store(storage cache.Storage, conf *models.Configuration) gin.HandlerFunc {
 				c.AbortWithError(http.StatusInternalServerError, errors.New("#S0003"))
 				return
 			}
+			logg.Debug("found ", len(apiPath), " apiPath for user ", account.Name)
 			if len(apiPath) == 0 {
 				logg.Error("NO API FOUND")
 				c.AbortWithError(http.StatusForbidden, errors.New("#S0004"))
@@ -70,6 +74,37 @@ func Store(storage cache.Storage, conf *models.Configuration) gin.HandlerFunc {
 		if routeType == "internal" {
 			w, _ := c.MustGet("wksid").(string)
 			wksid, _ := strconv.Atoi(w)
+			workers, err := models.GetWorkers(storage, conf)
+			var worker models.EzbWorkers
+			if err != nil {
+				logg.Error(err)
+				c.AbortWithError(http.StatusInternalServerError, errors.New("#S0005"))
+				return
+			}
+			for _, w := range workers {
+				if w.ID == wksid {
+					worker = w
+					break
+				}
+			}
+			if worker.ID == 0 {
+				logg.Error("NO WORKER FOUND")
+				c.AbortWithError(http.StatusForbidden, errors.New("#S0006"))
+				return
+			}
+			trace.Worker = worker.Name
+			trace.Controller = "internal"
+			c.Set("trace", trace)
+			c.Set("worker", worker)
+		}
+		if routeType == "tasks" {
+			tasksID := c.MustGet("tasksid").(string)
+			wksid, err := strconv.Atoi(tasksID[0:4])
+			if err != nil {
+				logg.Error("bad worker id")
+				c.AbortWithError(http.StatusBadRequest, errors.New("#S0007"))
+				return
+			}
 			workers, err := models.GetWorkers(storage, conf)
 			var worker models.EzbWorkers
 			if err != nil {